refactor(analysis): extract content function spans
